requirements.txt
langchain==0.0.335
langgraph==0.0.19
openai==1.3.7
python-dotenv==1.0.0
reportlab==4.0.4
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.13
colorama==0.4.6
gTTS==2.3.2
playsound==1.3.0
pygame==2.5.2
pocketsphinx==5.0.2
